adjusted all files, and improved layout
